Russian localization update
Hi Morse,
First of all I love your mod very much :)
I was leading the Russian localization team for KSP 1.3 and we had some standards we followed during localization. I updated the current Russian localization of your mod to the standards of the game, including grammar and terminology, this will set it in line with the language style we used. Btw, the original Russian localization is very good. Who made it? :)